Backend/requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.13
djangorestframework==3.15.1
djongo==1.3.6
dnspython==2.6.1
idna==3.7
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pymongo==4.7.3
requests==2.32.3
sqlparse==0.2.4
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
